Add WarehousePathPlanner class for robot path planning
